Feature 1.4.11 wc bms1530

See merge request bms/bms!250

+++ bms-site/src/main/java/com/jiuyescm/bms/base/file/web/FileExportTaskController.java
@@ -138,10 +138,12 @@ public class FileExportTaskController {
 			return MessageConstant.DELETE_INFO_NULL_MSG;
 		}
 		try {
-			File file = new File(entity.getFilePath());
-			if (file.exists()) {
-				file.delete();
-			}
+		    if (StringUtils.isNotBlank(entity.getFilePath())) {
+		        File file = new File(entity.getFilePath());
+	            if (file.exists()) {
+	                file.delete();
+	            }
+            }
 			FileExportTaskEntity delEntity = new FileExportTaskEntity();
 			delEntity.setTaskId(entity.getTaskId());
 			delEntity.setDelFlag("1");
